# services/base/docs/swagger.yaml
definitions:
  dto.AuthTokensDTO:
    properties:
      access_token:
        type: string
      email:
        type: string
      id: {}
      refresh_token:
        type: string
    type: object
...
  dto.EmailVerificationDTI:
    properties:
      token_hash:
        maxLength: 6
        minLength: 6
        type: string
    required:
    - token_hash
    type: object
  dto.ForgotPasswordDTI:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.RefreshTokenDTI:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterUserDTI:
    properties:
      email:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  dto.RequestEmailCodeDTI:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.ResetPasswordDTI:
    properties:
      confirm_password:
        type: string
      password:
        maxLength: 64
        minLength: 8
        type: string
      token:
        type: string
    required:
    - confirm_password
    - password
    - token
    type: object
...
  dto.UserExistDTI:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  dto.VerifyEmailCodeDTI:
    properties:
      code:
        type: string
      email:
        type: string
    required:
    - code
    - email
    type: object
...
  exception.ExceptionError:
    properties:
      error_subcode:
        type: integer
      message:
        type: string
      type:
        $ref: '#/definitions/exception.errorType'
    type: object
  exception.ExceptionErrors:
    properties:
      errItems:
        items:
          $ref: '#/definitions/exception.ExceptionError'
        type: array
      globalMessage:
        type: string
      httpStatusCode:
        type: integer
    type: object
  exception.errorType:
    enum:
    - UnknownError
    - BadRequest
    - InvalidUUID
    - NotFound
    - Unauthorized
    - ValidationError
    - JobError
    - ExternalServiceError
    - DatasourceError
    type: string
    x-enum-varnames:
    - ERROR_TYPE_UNKNOWN_ERROR
    - ERROR_TYPE_BAD_REQUEST
    - ERROR_TYPE_INVALID_UUID
    - ERROR_TYPE_NOT_FOUND
    - ERROR_TYPE_UNAUTHORIZED
    - ERROR_TYPE_VALIDATION_ERROR
    - ERROR_TYPE_JOB_ERROR
    - ERROR_TYPE_EXTERNAL_SERVICE_ERROR
    - ERROR_TYPE_DATASOURCE_ERROR
...
  response.CommonResponse:
    properties:
      data: {}
      errors:
        $ref: '#/definitions/exception.ExceptionErrors'
      request_id:
        type: string
      response_code:
        type: integer
      response_message:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/authentication/email-code/request:
    post:
      consumes:
      - application/json
      description: Sends a 6-character one-time code to the given email. Always returns
        200 to avoid leaking whether an email is registered. Code expires in 10 minutes.
      parameters:
      - description: Email address
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RequestEmailCodeDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Request a login code
      tags:
      - Authentication
  /api/v1/authentication/email-code/verify:
    post:
      consumes:
      - application/json
      description: Validates the one-time code sent to the user's email and returns
        an access + refresh token pair.
      parameters:
      - description: Email and code
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.VerifyEmailCodeDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.CommonResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.AuthTokensDTO'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Login with email code
      tags:
      - Authentication
  /api/v1/authentication/email/exist:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UserExistDTI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Check if email exists
      tags:
      - Authentication
  /api/v1/authentication/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ForgotPasswordDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Request password reset
      tags:
      - Authentication
...
  /api/v1/authentication/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Log in a user
      tags:
      - Authentication
  /api/v1/authentication/logout:
    post:
      consumes:
      - application/json
      description: Revokes the refresh token so it cannot be used again.
      parameters:
      - description: Refresh token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Logout
      tags:
      - Authentication
  /api/v1/authentication/refresh:
    post:
      consumes:
      - application/json
      description: Exchanges a valid refresh token for a new access + refresh token
        pair. The old refresh token is invalidated.
      parameters:
      - description: Refresh token
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshTokenDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Refresh access token
      tags:
      - Authentication
  /api/v1/authentication/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Registration data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterUserDTI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Register a new user
      tags:
      - Authentication
  /api/v1/authentication/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Reset data
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.ResetPasswordDTI'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Reset user password
      tags:
      - Authentication
...
  /api/v1/verification/email:
    post:
      consumes:
      - application/json
      description: Verifies a user's email address using the 6-digit code sent to
        their inbox
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Verification code
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.EmailVerificationDTI'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Verify email address
      tags:
      - Verification
  /api/v1/verification/email/resend:
    get:
      description: Sends a new verification code to the authenticated user's email
        address
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.CommonResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.CommonResponse'
      summary: Resend verification email
      tags:
      - Verification
...
swagger: "2.0"